Adjust config as code tests and fix indentation
