+++ src/main/java/com/microsoft/jenkins/azuread/AzureSecurityRealm.java
@@ -75,7 +75,6 @@ import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -128,7 +127,6 @@ public class AzureSecurityRealm extends SecurityRealm {
     private Cache<String, AzureAdUser> caches;
 
     private Secret clientId;
-
     private Secret clientSecret;
     private Secret clientCertificate;
     private Secret tenant;
@@ -185,7 +183,6 @@ public class AzureSecurityRealm extends SecurityRealm {
                 .httpClient(HttpClientRetriever.get())
                 .build();
     }
-
     public boolean isPromptAccount() {
         return promptAccount;
     }
