+++ src/main/java/com/microsoft/jenkins/azuread/GraphClientCacheKey.java
@@ -40,7 +40,6 @@ class GraphClientCacheKey {
     public String getClientId() {
         return clientId;
     }
-
     public String getClientSecret() {
         return clientSecret;
     }
@@ -53,9 +52,7 @@ class GraphClientCacheKey {
     public String getTenantId() {
         return tenantId;
     }
-
     public String getAzureEnvironmentName() {
         return azureEnvironmentName;
     }
-
 }
